+++ crates/algod_client/src/models/get_supply200_response.rs
@@ -17,13 +17,13 @@ use serde::{Deserialize, Serialize};
 pub struct GetSupply200Response {
     /// Round
     #[serde(rename = "current_round")]
-    pub current_round: i32,
+    pub current_round: u64,
     /// OnlineMoney
     #[serde(rename = "online-money")]
-    pub online_money: i32,
+    pub online_money: u64,
     /// TotalMoney
     #[serde(rename = "total-money")]
-    pub total_money: i32,
+    pub total_money: u64,
 }
 
 impl AlgorandMsgpack for GetSupply200Response {
@@ -32,7 +32,7 @@ impl AlgorandMsgpack for GetSupply200Response {
 
 impl GetSupply200Response {
     /// Constructor for GetSupply200Response
-    pub fn new(current_round: i32, online_money: i32, total_money: i32) -> GetSupply200Response {
+    pub fn new(current_round: u64, online_money: u64, total_money: u64) -> GetSupply200Response {
         GetSupply200Response {
             current_round,
             online_money,
